made code more DRY, used props to test dummy data
